% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{diff_words}
\alias{diff_words}
\title{Compute word-level diff using Google's diff-match-patch}
\usage{
diff_words(text1, text2)
}
\arguments{
\item{text1, }{text2 Strings to compare}
}
\value{
List of diff opcodes and tokens
}
\description{
Compute word-level diff using Google's diff-match-patch
}
