% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{get_metadata_across_commits}
\alias{get_metadata_across_commits}
\title{Extract LaTeX metadata (title and abstract) across commits}
\usage{
get_metadata_across_commits(owner, repo, path, commits)
}
\arguments{
\item{owner, }{repo, path - GitHub info}

\item{commits}{- list of commit objects (from gh::gh)}
}
\value{
named list of parsed data per commit SHA
}
\description{
Extract LaTeX metadata (title and abstract) across commits
}
