% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{get_raw_tex_snippets}
\alias{get_raw_tex_snippets}
\title{Get raw text content of a LaTeX file at each commit}
\usage{
get_raw_tex_snippets(owner, repo, path, commits)
}
\arguments{
\item{owner, }{repo, path - GitHub info}

\item{commits}{- list of commit objects}
}
\value{
named list of first 10 lines of raw .tex at each commit
}
\description{
Get raw text content of a LaTeX file at each commit
}
