% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{render_diff_html}
\alias{render_diff_html}
\title{Render a diff as HTML with <ins> and <del> tags}
\usage{
render_diff_html(decoded_diffs)
}
\arguments{
\item{decoded_diffs}{A list from diff_words()}
}
\value{
HTML fragment (as htmltools::HTML)
}
\description{
Render a diff as HTML with <ins> and <del> tags
}
