% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{split_words}
\alias{split_words}
\title{Tokenize LaTeX text into words and symbols}
\usage{
split_words(text)
}
\arguments{
\item{text}{Character string}
}
\value{
Character vector of tokens
}
\description{
Tokenize LaTeX text into words and symbols
}
